+++ src/grafv4/mod.rs
@@ -1,5 +1,5 @@
-pub(crate) mod count_mode_selector;
+pub(crate) mod count_mode;
 mod graf_zahl;
 mod io_reader;
 mod countable;
-mod count_modes;
+mod count_mode_lib;
