+++ src/main.rs
@@ -9,7 +9,7 @@ use structopt::StructOpt;
 
 use crate::grafzahl::ignore_checker::{init_empty_list, init_ignore_list};
 use crate::grafzahl::languages::{import_languages, Language};
-use crate::grafv4::count_mode_selector::{CountMode, execute_count_mode};
+use crate::grafv4::count_mode::{CountMode, execute_count_mode};
 
 mod grafzahl;
 mod grafv4;
